Refactor: Separate Dealer, Player, Participant Feat: Add remove(card:) to Dealer Feat: Add receive(card:) to Player
