+++ PokerGameApp/PokerGameApp/PokerGame.swift
@@ -8,10 +8,11 @@
 import Foundation
 
 struct PokerGame {
-    var dealer: Dealer? // run 하면 Dealer를 지정합니다.
-    var players = [Player]() // run 하면 players를 추가해줍니다.
-    static var totalCards = CardDeck() // 전체 카드
-    static var cardStud: Int = Stud.five.rawValue // default는 5카드 스터드
+    private var dealer: Dealer? // run 하면 Dealer를 지정합니다.
+    private var players = [Player]() // run 하면 players를 추가해줍니다.
+    private var totalCards = CardDeck() // 전체 카드
+    private var receivedCard: Card? // 딜러로부터 받은 카드
+    private var cardStud: Int = Stud.five.rawValue // default는 5카드 스터드
     
     
     enum Stud: Int {
@@ -22,30 +23,30 @@ struct PokerGame {
     
     // 게임을 실행합니다.
     mutating func run() {
-        let playersCount = Int.random(in: 1...4) // 플레이어는 1 ~ 4명 (Int.random(in: 1...4))
-        let playerNames = makeRandomName(count: playersCount) // 참가자 이름 배열 생성
-        
-        // 플레이어 추가
-        for name in playerNames {
-            self.players.append(Player(name: name))
-        }
-        
-        dealer = Dealer()
-        
-        var totalCardCount = PokerGame.totalCards.count
-        PokerGame.totalCards.shuffle()
-        
-        // 배분할 카드 수 < 참여자 + 딜러 수이거나, 전체 카드 수가 0이면 종료
-        while !(totalCardCount < playersCount + 1) && totalCardCount != 0 {
-            guard let distributedCardCount = dealer?.distributeCard(to: &players) else {
-                break
-            }
-            totalCardCount -= distributedCardCount
-            
-            if dealer?.cards.count == PokerGame.cardStud { // 카드 5장 분배하면 종료..
-                break
-            }
-        }
+//        let playersCount = Int.random(in: 1...4) // 플레이어는 1 ~ 4명 (Int.random(in: 1...4))
+//        let playerNames = makeRandomName(count: playersCount) // 참가자 이름 배열 생성
+//        
+//        // 플레이어 추가
+//        for name in playerNames {
+//            self.players.append(Player(name: name))
+//        }
+//        
+//        dealer = Dealer()
+//        
+//        var totalCardCount = PokerGame.totalCards.count
+//        PokerGame.totalCards.shuffle()
+//        
+//        // 배분할 카드 수 < 참여자 + 딜러 수이거나, 전체 카드 수가 0이면 종료
+//        while !(totalCardCount < playersCount + 1) && totalCardCount != 0 {
+//            guard let distributedCardCount = dealer?.distributeCard(to: &players) else {
+//                break
+//            }
+//            totalCardCount -= distributedCardCount
+//            
+//            if dealer?.cards.count == PokerGame.cardStud { // 카드 5장 분배하면 종료..
+//                break
+//            }
+//        }
         
         
     }
@@ -66,46 +67,6 @@ struct PokerGame {
         }
         return players
     }
-    
-    
-    struct Dealer: Participant, CustomStringConvertible {
-        var description: String {
-            return "😎Dealer: \(self.cards)"
-        }
-        
-        var cards = [Card]()
-        
-        // 자신을 포함한 플레이어에게 카드를 나눠주고 나눠준 총 카드 수를 리턴합니다.
-        mutating func distributeCard(to players: inout [Player]) -> Int {
-            guard let pickedCard = totalCards.removeOne() else {
-                return -1
-            }
-            cards.append(pickedCard)
-            
-            for player in 0..<players.count {
-                guard let pickedCard = totalCards.removeOne() else {
-                    return -1
-                }
-                players[player].cards.append(pickedCard)
-            }
-            return cards.count
-        }
-    }
-    
-    
-    struct Player: Participant, CustomStringConvertible {
-        var description: String {
-            return "👤User(\(name)): \(self.cards)"
-        }
-        
-        var name: String
-        var cards = [Card]()
-    }
-}
-
-
-protocol Participant {
-    var cards: [Card] { get set } // 딜러, 플레이어는 모두 카드 갖고 있음. Card 타입을 배열로 갖고 있음
 }
